+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change log
 
+## 0.9.0 (2024-05-XX)
+
+### Features
+
+- popup: Added a Bandcamp search form on the Warning message tab
+- popup: Added link to the changelog file on "About" tab
+- core: Added `externalContentLoaded` event to `external-content` custom HTML tag element
+
+### Fixed
+
+- popup: Fixed renaming "Storage" tab to "History" tab
+
 ## 0.8.0 (2024-04-23)
 
 ### Features